+++ build-logic/plugins/src/main/kotlin/uk/gov/onelogin/criorchestrator/imposter-test-config.gradle.kts
@@ -1,0 +1,10 @@
+package uk.gov.onelogin.criorchestrator
+
+import org.gradle.accessors.dm.LibrariesForLibs
+import uk.gov.onelogin.criorchestrator.extensions.imposterTestDependencies
+
+val libs = the<LibrariesForLibs>()
+
+dependencies {
+    imposterTestDependencies(libs)
+}
