+++ libraries/testing/src/testFixtures/kotlin/uk/gov/onelogin/criorchestrator/libraries/testing/networking/CreateTestHttpClient.kt
@@ -1,4 +1,4 @@
-package uk.gov.onelogin.criorchestrator.features.session.internal
+package uk.gov.onelogin.criorchestrator.libraries.testing.networking
 
 import uk.gov.android.network.auth.AuthenticationProvider
 import uk.gov.android.network.auth.AuthenticationResponse
